pyproject.toml
[tool.poetry]
name = "vyxalbot2"
version = "0.1.0"
description = ""
authors = ["GingerIndustries <75683114+gingerindustries@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gidgethub = "^5.2.1"
sechat = { git = "https://github.com/GingerIndustries/sechat.git", branch = "v2" }
cachetools = "^5.3.1"
tomli = "^2.0.1"
tinydb = "^4.7.1"
pyjwt = "^2.7.0"
python-dateutil = "^2.8.2"
websockets = "^10.4"
uwuivy = { git = "https://github.com/Vyxal/uwuivy.git", branch = "v0.1.0" }
pyyaml = "^6.0"
odmantic = "^0.9.2"
blinker = "^1.7.0"
pydantic = "<2.0"
discord-py = "^2.3.2"

[tool.poetry.group.dev.dependencies]
black = {version = "^23.9.1", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
